moved member functions declarations to new file and created new folder
